Some format corrections and more organization (focused in route files and controllers)
