+++ controllers/middleware.js
@@ -1,3 +1,5 @@
+const jwt = require('jsonwebtoken');
+
 exports.checkToken = (req, res, next) => {
     const header = req.headers['authorization'];
 
@@ -6,7 +8,17 @@ exports.checkToken = (req, res, next) => {
         const token = bearer[1];
 
         req.token = token;
-        next();
+
+        jwt.verify(req.token, process.env.JWT_SECRET, (err, authorizedData) => {
+            if (err) {
+                res.status(403).json({ 
+                    message: 'Autentication failed.', 
+                    error: err.message 
+                });
+            } else {
+                next();
+            }
+        });
     } else {
         res.status(403).json({ message: 'Authentication credentials were not provided' });
     }
