+++ routes/index.js
@@ -4,10 +4,13 @@ const router = express.Router();
 const controller = require('../controllers/controller');
 const middleware = require('../controllers/middleware');
 const authController = require('../controllers/authController');
+const userController = require('../controllers/userController');
+
+const usersRouter = require('./users');
 
 /* GET home page. */
 router.get('/', middleware.checkToken, controller.index_get);
 router.post('/login', authController.login);
-router.post('/signup', middleware.checkToken, controller.user_create_post);
+router.get('/users', usersRouter);
 
 module.exports = router;
